@@ -1,59 +1,55 @@
-﻿using System;
-using System.Diagnostics;
-using System.IO;
-using System.Threading;
+﻿using System.Diagnostics;
 using Xunit;
 
-namespace Lunar.Tests
+namespace Lunar.Tests;
+
+[Collection("LibraryMapper Tests")]
+public sealed class X86Tests : IDisposable
 {
-    [Collection("LibraryMapper Tests")]
-    public sealed class X86Tests : IDisposable
+    private readonly string _testBinaryDirectoryPath;
+    private readonly Process _process;
+
+    public X86Tests()
     {
-        private readonly string _testBinaryDirectoryPath;
-        private readonly Process _process;
+        _testBinaryDirectoryPath = Path.GetFullPath(@"..\..\..\TestBinaries\bin\x86\Release");
+        _process = new Process { StartInfo = { FileName = Path.Combine(_testBinaryDirectoryPath, "Target.exe"), UseShellExecute = true, WindowStyle = ProcessWindowStyle.Hidden } };
+        _process.Start();
 
-        public X86Tests()
-        {
-            _testBinaryDirectoryPath = Path.GetFullPath(@"..\..\..\TestBinaries\bin\x86\Release");
-            _process = new Process {StartInfo = {FileName = Path.Combine(_testBinaryDirectoryPath, "Target.exe"), UseShellExecute = true, WindowStyle = ProcessWindowStyle.Hidden}};
-            _process.Start();
+        // Wait an arbitrary amount of time (10 milliseconds) for the process to initialise
 
-            // Wait an arbitrary amount of time (10 milliseconds) for the process to initialise
+        Thread.Sleep(10);
+    }
 
-            Thread.Sleep(10);
-        }
+    public void Dispose()
+    {
+        _process.Kill();
+        _process.Dispose();
+    }
 
-        public void Dispose()
-        {
-            _process.Kill();
-            _process.Dispose();
-        }
+    [Theory]
+    [InlineData("Basic.dll")]
+    [InlineData("Exception.dll")]
+    [InlineData("StaticTls.dll")]
+    [InlineData("TlsCallback.dll")]
+    public void TestMap(string dllName)
+    {
+        var libraryMapper = new LibraryMapper(_process, Path.Combine(_testBinaryDirectoryPath, dllName));
+        libraryMapper.MapLibrary();
 
-        [Theory]
-        [InlineData("Basic.dll")]
-        [InlineData("Exception.dll")]
-        [InlineData("StaticTls.dll")]
-        [InlineData("TlsCallback.dll")]
-        public void TestMap(string dllName)
-        {
-            var libraryMapper = new LibraryMapper(_process, Path.Combine(_testBinaryDirectoryPath, dllName));
-            libraryMapper.MapLibrary();
+        Assert.NotEqual(libraryMapper.DllBaseAddress, IntPtr.Zero);
+    }
 
-            Assert.NotEqual(libraryMapper.DllBaseAddress, IntPtr.Zero);
-        }
+    [Theory]
+    [InlineData("Basic.dll")]
+    [InlineData("Exception.dll")]
+    [InlineData("StaticTls.dll")]
+    [InlineData("TlsCallback.dll")]
+    public void TestUnmap(string dllName)
+    {
+        var libraryMapper = new LibraryMapper(_process, Path.Combine(_testBinaryDirectoryPath, dllName));
+        libraryMapper.MapLibrary();
+        libraryMapper.UnmapLibrary();
 
-        [Theory]
-        [InlineData("Basic.dll")]
-        [InlineData("Exception.dll")]
-        [InlineData("StaticTls.dll")]
-        [InlineData("TlsCallback.dll")]
-        public void TestUnmap(string dllName)
-        {
-            var libraryMapper = new LibraryMapper(_process, Path.Combine(_testBinaryDirectoryPath, dllName));
-            libraryMapper.MapLibrary();
-            libraryMapper.UnmapLibrary();
-
-            Assert.Equal(libraryMapper.DllBaseAddress, IntPtr.Zero);
-        }
+        Assert.Equal(libraryMapper.DllBaseAddress, IntPtr.Zero);
     }
 }
