@@ -1,14 +1,10 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Diagnostics;
-using System.IO;
-using System.Linq;
+﻿using System.Diagnostics;
+using System.Numerics;
 using System.Reflection.PortableExecutable;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Security.Cryptography;
 using System.Text;
-using System.Threading.Tasks;
 using Lunar.Extensions;
 using Lunar.FileResolution;
 using Lunar.Native;
@@ -18,288 +14,290 @@ using Lunar.PortableExecutable;
 using Lunar.Remote;
 using Lunar.Utilities;
 
-namespace Lunar
+namespace Lunar;
+
+/// <summary>
+/// Provides the functionality to map a DLL from disk or memory into a process
+/// </summary>
+public sealed class LibraryMapper
 {
     /// <summary>
-    /// Provides the functionality to map a DLL from disk or memory into a process
+    /// The base address of the DLL in the process
     /// </summary>
-    public sealed class LibraryMapper
+    public IntPtr DllBaseAddress { get; private set; }
+
+    private readonly Memory<byte> _dllBytes;
+    private readonly FileResolver _fileResolver;
+    private readonly MappingFlags _mappingFlags;
+    private readonly PeImage _peImage;
+    private readonly ProcessContext _processContext;
+    private (IntPtr EntryAddress, int Index, bool ModifiedBitmap) _tlsData;
+
+    /// <summary>
+    /// Initialises an instances of the <see cref="LibraryMapper"/> class with the functionality to map a DLL from memory into a process
+    /// </summary>
+    public LibraryMapper(Process process, Memory<byte> dllBytes, MappingFlags mappingFlags = MappingFlags.None)
     {
-        /// <summary>
-        /// The base address of the DLL in the process
-        /// </summary>
-        public IntPtr DllBaseAddress { get; private set; }
-
-        private readonly Memory<byte> _dllBytes;
-        private readonly FileResolver _fileResolver;
-        private readonly MappingFlags _mappingFlags;
-        private readonly PeImage _peImage;
-        private readonly ProcessContext _processContext;
-        private (IntPtr EntryAddress, int Index, bool ModifiedBitmap) _tlsData;
-
-        /// <summary>
-        /// Initialises an instances of the <see cref="LibraryMapper"/> class with the functionality to map a DLL from memory into a process
-        /// </summary>
-        public LibraryMapper(Process process, Memory<byte> dllBytes, MappingFlags mappingFlags = MappingFlags.None)
+        if (process.HasExited)
         {
-            if (process.HasExited)
-            {
-                throw new ArgumentException("The provided process is not currently running");
-            }
-
-            if (dllBytes.IsEmpty)
-            {
-                throw new ArgumentException("The provided DLL bytes were empty");
-            }
-
-            if (!Environment.Is64BitProcess && process.GetArchitecture() == Architecture.X64)
-            {
-                throw new NotSupportedException("The provided process cannot be mapped into from an x86 build");
-            }
-
-            _dllBytes = dllBytes.ToArray();
-            _fileResolver = new FileResolver(process, null);
-            _mappingFlags = mappingFlags;
-            _peImage = new PeImage(dllBytes);
-            _processContext = new ProcessContext(process);
+            throw new ArgumentException("The provided process is not currently running");
         }
 
-        /// <summary>
-        /// Initialises an instances of the <see cref="LibraryMapper"/> class with the functionality to map a DLL from disk into a process
-        /// </summary>
-        public LibraryMapper(Process process, string dllFilePath, MappingFlags mappingFlags = MappingFlags.None)
+        if (dllBytes.IsEmpty)
         {
-            if (process.HasExited)
-            {
-                throw new ArgumentException("The provided process is not currently running");
-            }
-
-            if (!File.Exists(dllFilePath))
-            {
-                throw new ArgumentException("The provided file path did not point to a valid file");
-            }
-
-            if (!Environment.Is64BitProcess && process.GetArchitecture() == Architecture.X64)
-            {
-                throw new NotSupportedException("The provided process cannot be mapped into from an x86 build");
-            }
-
-            _dllBytes = File.ReadAllBytes(dllFilePath);
-            _fileResolver = new FileResolver(process, Path.GetDirectoryName(dllFilePath));
-            _mappingFlags = mappingFlags;
-            _peImage = new PeImage(File.ReadAllBytes(dllFilePath));
-            _processContext = new ProcessContext(process);
+            throw new ArgumentException("The provided DLL bytes were empty");
         }
 
-        /// <summary>
-        /// Maps the DLL into the process
-        /// </summary>
-        public void MapLibrary()
+        if (!Environment.Is64BitProcess && process.GetArchitecture() == Architecture.X64)
         {
-            if (DllBaseAddress != IntPtr.Zero)
+            throw new NotSupportedException("The provided process cannot be mapped into from an x86 build");
+        }
+
+        _dllBytes = dllBytes.ToArray();
+        _fileResolver = new FileResolver(process, null);
+        _mappingFlags = mappingFlags;
+        _peImage = new PeImage(dllBytes);
+        _processContext = new ProcessContext(process);
+    }
+
+    /// <summary>
+    /// Initialises an instances of the <see cref="LibraryMapper"/> class with the functionality to map a DLL from disk into a process
+    /// </summary>
+    public LibraryMapper(Process process, string dllFilePath, MappingFlags mappingFlags = MappingFlags.None)
+    {
+        if (process.HasExited)
+        {
+            throw new ArgumentException("The provided process is not currently running");
+        }
+
+        if (!File.Exists(dllFilePath))
+        {
+            throw new ArgumentException("The provided file path did not point to a valid file");
+        }
+
+        if (!Environment.Is64BitProcess && process.GetArchitecture() == Architecture.X64)
+        {
+            throw new NotSupportedException("The provided process cannot be mapped into from an x86 build");
+        }
+
+        _dllBytes = File.ReadAllBytes(dllFilePath);
+        _fileResolver = new FileResolver(process, Path.GetDirectoryName(dllFilePath));
+        _mappingFlags = mappingFlags;
+        _peImage = new PeImage(File.ReadAllBytes(dllFilePath));
+        _processContext = new ProcessContext(process);
+    }
+
+    /// <summary>
+    /// Maps the DLL into the process
+    /// </summary>
+    public void MapLibrary()
+    {
+        if (DllBaseAddress != IntPtr.Zero)
+        {
+            return;
+        }
+
+        var cleanupStack = new Stack<Action>();
+
+        try
+        {
+            DllBaseAddress = _processContext.Process.AllocateBuffer(_peImage.Headers.PEHeader!.SizeOfImage, ProtectionType.ReadOnly);
+            cleanupStack.Push(() => _processContext.Process.FreeBuffer(DllBaseAddress));
+
+            LoadDependencies();
+            cleanupStack.Push(FreeDependencies);
+
+            BuildImportAddressTable();
+            RelocateImage();
+
+            if (!_mappingFlags.HasFlag(MappingFlags.DiscardHeaders))
             {
-                return;
+                MapHeaders();
             }
 
-            var cleanupStack = new Stack<Action>();
+            MapSections();
+            InitialiseControlFlowGuard();
+            InitialiseSecurityCookie();
 
-            try
+            InsertExceptionHandlers();
+            cleanupStack.Push(RemoveExceptionHandlers);
+
+            ReserveTlsIndex();
+            cleanupStack.Push(ReleaseTlsIndex);
+
+            InitialiseTlsData();
+            cleanupStack.Push(FreeTlsEntry);
+
+            if (!_mappingFlags.HasFlag(MappingFlags.SkipInitialisationRoutines))
             {
-                DllBaseAddress = _processContext.Process.AllocateBuffer(_peImage.Headers.PEHeader!.SizeOfImage, ProtectionType.ReadOnly);
-                cleanupStack.Push(() => _processContext.Process.FreeBuffer(DllBaseAddress));
-
-                LoadDependencies();
-                cleanupStack.Push(FreeDependencies);
-
-                BuildImportAddressTable();
-                RelocateImage();
-
-                if (!_mappingFlags.HasFlag(MappingFlags.DiscardHeaders))
-                {
-                    MapHeaders();
-                }
-
-                MapSections();
-                InitialiseControlFlowGuard();
-                InitialiseSecurityCookie();
-
-                InsertExceptionHandlers();
-                cleanupStack.Push(RemoveExceptionHandlers);
-
-                ReserveTlsIndex();
-                cleanupStack.Push(ReleaseTlsIndex);
-
-                InitialiseTlsData();
-                cleanupStack.Push(FreeTlsEntry);
-
-                if (!_mappingFlags.HasFlag(MappingFlags.SkipInitialisationRoutines))
-                {
-                    CallInitialisationRoutines(DllReason.ProcessAttach);
-                }
-            }
-
-            catch
-            {
-                while (cleanupStack.TryPop(out var cleanupRoutine))
-                {
-                    Executor.IgnoreExceptions(cleanupRoutine);
-                }
-
-                DllBaseAddress = IntPtr.Zero;
-                throw;
+                CallInitialisationRoutines(DllReason.ProcessAttach);
             }
         }
 
-        /// <summary>
-        /// Unmaps the DLL from the process
-        /// </summary>
-        public void UnmapLibrary()
+        catch
         {
-            if (DllBaseAddress == IntPtr.Zero)
+            while (cleanupStack.TryPop(out var cleanupRoutine))
             {
-                return;
+                Executor.IgnoreExceptions(cleanupRoutine);
             }
 
-            var topLevelException = default(Exception);
+            DllBaseAddress = IntPtr.Zero;
+            throw;
+        }
+    }
 
-            try
-            {
-                if (!_mappingFlags.HasFlag(MappingFlags.SkipInitialisationRoutines))
-                {
-                    CallInitialisationRoutines(DllReason.ProcessDetach);
-                }
-            }
+    /// <summary>
+    /// Unmaps the DLL from the process
+    /// </summary>
+    public void UnmapLibrary()
+    {
+        if (DllBaseAddress == IntPtr.Zero)
+        {
+            return;
+        }
 
-            catch (Exception exception)
-            {
-                topLevelException ??= exception;
-            }
+        var topLevelException = default(Exception);
 
-            try
+        try
+        {
+            if (!_mappingFlags.HasFlag(MappingFlags.SkipInitialisationRoutines))
             {
-                FreeTlsEntry();
-            }
-
-            catch (Exception exception)
-            {
-                topLevelException ??= exception;
-            }
-
-            try
-            {
-                ReleaseTlsIndex();
-            }
-
-            catch (Exception exception)
-            {
-                topLevelException ??= exception;
-            }
-
-            try
-            {
-                RemoveExceptionHandlers();
-            }
-
-            catch (Exception exception)
-            {
-                topLevelException ??= exception;
-            }
-
-            try
-            {
-                FreeDependencies();
-            }
-
-            catch (Exception exception)
-            {
-                topLevelException ??= exception;
-            }
-
-            try
-            {
-                _processContext.Process.FreeBuffer(DllBaseAddress);
-            }
-
-            catch (Exception exception)
-            {
-                topLevelException ??= exception;
-            }
-
-            finally
-            {
-                DllBaseAddress = IntPtr.Zero;
-            }
-
-            if (topLevelException is not null)
-            {
-                throw topLevelException;
+                CallInitialisationRoutines(DllReason.ProcessDetach);
             }
         }
 
-        private void BuildImportAddressTable()
+        catch (Exception exception)
         {
-            Parallel.ForEach(_peImage.ImportDirectory.GetImportDescriptors(), importDescriptor =>
-            {
-                foreach (var (functionName, functionOffset, functionOrdinal) in importDescriptor.Functions)
-                {
-                    // Write the function address into the import address table
-
-                    var functionAddress = functionName is null ? _processContext.GetFunctionAddress(importDescriptor.Name, functionOrdinal) : _processContext.GetFunctionAddress(importDescriptor.Name, functionName);
-                    MemoryMarshal.Write(_dllBytes.Span[functionOffset..], ref functionAddress);
-                }
-            });
+            topLevelException ??= exception;
         }
 
-        private void CallInitialisationRoutines(DllReason reason)
+        try
         {
-            // Call the entry point of any TLS callbacks
+            FreeTlsEntry();
+        }
 
-            foreach (var callbackAddress in _peImage.TlsDirectory.GetTlsCallbacks().Select(callBack => DllBaseAddress + callBack.RelativeAddress))
+        catch (Exception exception)
+        {
+            topLevelException ??= exception;
+        }
+
+        try
+        {
+            ReleaseTlsIndex();
+        }
+
+        catch (Exception exception)
+        {
+            topLevelException ??= exception;
+        }
+
+        try
+        {
+            RemoveExceptionHandlers();
+        }
+
+        catch (Exception exception)
+        {
+            topLevelException ??= exception;
+        }
+
+        try
+        {
+            FreeDependencies();
+        }
+
+        catch (Exception exception)
+        {
+            topLevelException ??= exception;
+        }
+
+        try
+        {
+            _processContext.Process.FreeBuffer(DllBaseAddress);
+        }
+
+        catch (Exception exception)
+        {
+            topLevelException ??= exception;
+        }
+
+        finally
+        {
+            DllBaseAddress = IntPtr.Zero;
+        }
+
+        if (topLevelException is not null)
+        {
+            throw topLevelException;
+        }
+    }
+
+    private void BuildImportAddressTable()
+    {
+        Parallel.ForEach(_peImage.ImportDirectory.GetImportDescriptors(), importDescriptor =>
+        {
+            foreach (var (functionName, functionOffset, functionOrdinal) in importDescriptor.Functions)
             {
-                _processContext.CallRoutine(callbackAddress, DllBaseAddress, reason, 0);
+                // Write the function address into the import address table
+
+                var functionAddress = functionName is null ? _processContext.GetFunctionAddress(importDescriptor.Name, functionOrdinal) : _processContext.GetFunctionAddress(importDescriptor.Name, functionName);
+                MemoryMarshal.Write(_dllBytes.Span[functionOffset..], ref functionAddress);
             }
+        });
+    }
 
-            if (_peImage.Headers.PEHeader!.AddressOfEntryPoint == 0)
+    private void CallInitialisationRoutines(DllReason reason)
+    {
+        // Call the entry point of any TLS callbacks
+
+        foreach (var callbackAddress in _peImage.TlsDirectory.GetTlsCallbacks().Select(callBack => DllBaseAddress + callBack.RelativeAddress))
+        {
+            _processContext.CallRoutine(callbackAddress, DllBaseAddress, reason, 0);
+        }
+
+        if (_peImage.Headers.PEHeader!.AddressOfEntryPoint == 0)
+        {
+            return;
+        }
+
+        // Call the DLL entry point
+
+        var entryPointAddress = DllBaseAddress + _peImage.Headers.PEHeader!.AddressOfEntryPoint;
+
+        if (!_processContext.CallRoutine<bool>(entryPointAddress, DllBaseAddress, reason, 0))
+        {
+            throw new ApplicationException($"Failed to call the DLL entry point with {reason:G}");
+        }
+    }
+
+    private void FreeDependencies()
+    {
+        foreach (var (_, dependencyName) in _peImage.ImportDirectory.GetImportDescriptors())
+        {
+            // Free the dependency using the Windows loader
+
+            var dependencyAddress = _processContext.GetModuleAddress(dependencyName);
+
+            if (!_processContext.CallRoutine<bool>(_processContext.GetFunctionAddress("kernel32.dll", "FreeLibrary"), dependencyAddress))
             {
-                return;
-            }
-
-            // Call the DLL entry point
-
-            var entryPointAddress = DllBaseAddress + _peImage.Headers.PEHeader!.AddressOfEntryPoint;
-
-            if (!_processContext.CallRoutine<bool>(entryPointAddress, DllBaseAddress, reason, 0))
-            {
-                throw new ApplicationException($"Failed to call the DLL entry point with {reason:G}");
+                throw new ApplicationException($"Failed to free the dependency {dependencyName} from the process");
             }
         }
 
-        private void FreeDependencies()
+        _processContext.ClearModuleCache();
+    }
+
+    private void FreeTlsEntry()
+    {
+        if (_peImage.Headers.PEHeader!.ThreadLocalStorageTableDirectory.Size == 0)
         {
-            foreach (var (_, dependencyName) in _peImage.ImportDirectory.GetImportDescriptors())
-            {
-                // Free the dependency using the Windows loader
-
-                var dependencyAddress = _processContext.GetModuleAddress(dependencyName);
-
-                if (!_processContext.CallRoutine<bool>(_processContext.GetFunctionAddress("kernel32.dll", "FreeLibrary"), dependencyAddress))
-                {
-                    throw new ApplicationException($"Failed to free the dependency {dependencyName} from the process");
-                }
-            }
-
-            _processContext.ClearModuleCache();
+            return;
         }
 
-        private void FreeTlsEntry()
+        _processContext.PebLock.Acquire();
+
+        try
         {
-            if (_peImage.Headers.PEHeader!.ThreadLocalStorageTableDirectory.Size == 0)
-            {
-                return;
-            }
-
-            using var pebLock = new SafePebLock(_processContext);
-
             if (_processContext.Architecture == Architecture.X86)
             {
                 // Read the TLS entry
@@ -310,13 +308,11 @@ namespace Lunar
 
                 var previousEntryAddress = UnsafeHelpers.WrapPointer(tlsEntry.EntryLinks.Blink);
                 var previousEntry = _processContext.Process.ReadStruct<ListEntry32>(previousEntryAddress);
-                previousEntry = new ListEntry32(tlsEntry.EntryLinks.Flink, previousEntry.Blink);
-                _processContext.Process.WriteStruct(previousEntryAddress, previousEntry);
+                _processContext.Process.WriteStruct(previousEntryAddress, previousEntry with { Flink = tlsEntry.EntryLinks.Flink });
 
                 var nextEntryAddress = UnsafeHelpers.WrapPointer(tlsEntry.EntryLinks.Flink);
                 var nextEntry = _processContext.Process.ReadStruct<ListEntry32>(nextEntryAddress);
-                nextEntry = new ListEntry32(nextEntry.Flink, tlsEntry.EntryLinks.Blink);
-                _processContext.Process.WriteStruct(nextEntryAddress, nextEntry);
+                _processContext.Process.WriteStruct(nextEntryAddress, nextEntry with { Blink = tlsEntry.EntryLinks.Blink });
             }
 
             else
@@ -329,126 +325,121 @@ namespace Lunar
 
                 var previousEntryAddress = UnsafeHelpers.WrapPointer(tlsEntry.EntryLinks.Blink);
                 var previousEntry = _processContext.Process.ReadStruct<ListEntry64>(previousEntryAddress);
-                previousEntry = new ListEntry64(tlsEntry.EntryLinks.Flink, previousEntry.Blink);
-                _processContext.Process.WriteStruct(previousEntryAddress, previousEntry);
+                _processContext.Process.WriteStruct(previousEntryAddress, previousEntry with { Flink = tlsEntry.EntryLinks.Flink });
 
                 var nextEntryAddress = UnsafeHelpers.WrapPointer(tlsEntry.EntryLinks.Flink);
                 var nextEntry = _processContext.Process.ReadStruct<ListEntry64>(nextEntryAddress);
-                nextEntry = new ListEntry64(nextEntry.Flink, tlsEntry.EntryLinks.Blink);
-                _processContext.Process.WriteStruct(nextEntryAddress, nextEntry);
+                _processContext.Process.WriteStruct(nextEntryAddress, nextEntry with { Blink = tlsEntry.EntryLinks.Blink });
             }
 
             // Free the TLS entry
 
-            _processContext.Process.FreeBuffer(_tlsData.EntryAddress);
+            _processContext.HeapManager.FreeBuffer(_tlsData.EntryAddress);
         }
 
-        private void InitialiseControlFlowGuard()
+        finally
         {
-            var loadConfigData = _peImage.LoadConfigDirectory.GetLoadConfigData();
+            _processContext.PebLock.Release();
+        }
+    }
 
-            if (loadConfigData is null || !loadConfigData.GuardFlags.HasFlag(GuardFlags.Instrumented))
-            {
-                return;
-            }
+    private void InitialiseControlFlowGuard()
+    {
+        var loadConfigData = _peImage.LoadConfigDirectory.GetLoadConfigData();
 
-            // Check if the process is using control flow guard
+        if (loadConfigData is null || !loadConfigData.GuardFlags.HasFlag(GuardFlags.Instrumented))
+        {
+            return;
+        }
 
-            if (!_processContext.CallRoutine<bool>(_processContext.GetFunctionAddress("ntdll.dll", "LdrControlFlowGuardEnforced")))
-            {
-                return;
-            }
+        // Check if the process is using control flow guard
 
-            // Check if the process is using export suppression
+        if (!_processContext.CallRoutine<bool>(_processContext.GetFunctionAddress("ntdll.dll", "LdrControlFlowGuardEnforced")))
+        {
+            return;
+        }
 
-            var usingExportSuppression = false;
+        // Check if the process is using export suppression
 
-            if (loadConfigData.GuardFlags.HasFlag(GuardFlags.ExportSuppressionInfoPresent))
-            {
-                usingExportSuppression = _processContext.CallRoutine<bool>(_processContext.GetNtdllSymbolAddress("LdrControlFlowGuardEnforcedWithExportSuppression"));
-            }
+        var usingExportSuppression = false;
 
-            // Get the address of the control flow guard functions
+        if (loadConfigData.GuardFlags.HasFlag(GuardFlags.ExportSuppressionInfoPresent))
+        {
+            usingExportSuppression = _processContext.CallRoutine<bool>(_processContext.GetNtdllSymbolAddress("LdrControlFlowGuardEnforcedWithExportSuppression"));
+        }
 
-            var checkFunctionName = "LdrpValidateUserCallTarget";
-            var dispatchFunctionName = "LdrpDispatchUserCallTarget";
+        var checkFunctionName = "LdrpValidateUserCallTarget";
+        var dispatchFunctionName = "LdrpDispatchUserCallTarget";
 
-            if (usingExportSuppression)
-            {
-                checkFunctionName = $"{checkFunctionName}ES";
-                dispatchFunctionName = $"{dispatchFunctionName}ES";
-            }
+        if (usingExportSuppression)
+        {
+            checkFunctionName = $"{checkFunctionName}ES";
+            dispatchFunctionName = $"{dispatchFunctionName}ES";
+        }
 
-            var checkFunctionAddress = _processContext.GetNtdllSymbolAddress(checkFunctionName);
-            var dispatchFunctionAddress = _peImage.Headers.PEHeader!.Magic == PEMagic.PE32 ? IntPtr.Zero : _processContext.GetNtdllSymbolAddress(dispatchFunctionName);
+        // Read the load config directory
+
+        var loadConfigDirectoryAddress = DllBaseAddress + _peImage.Headers.PEHeader!.LoadConfigTableDirectory.RelativeVirtualAddress;
+
+        if (_peImage.Headers.PEHeader!.Magic == PEMagic.PE32)
+        {
+            var loadConfigDirectory = _processContext.Process.ReadStruct<ImageLoadConfigDirectory32>(loadConfigDirectoryAddress);
 
             // Update the check function pointer
 
-            var loadConfigDirectoryAddress = DllBaseAddress + _peImage.Headers.PEHeader!.LoadConfigTableDirectory.RelativeVirtualAddress;
-
-            IntPtr checkFunctionUpdateAddress;
-
-            if (_peImage.Headers.PEHeader!.Magic == PEMagic.PE32)
-            {
-                checkFunctionUpdateAddress = loadConfigDirectoryAddress + Marshal.OffsetOf<ImageLoadConfigDirectory32>("GuardCFCheckFunctionPointer").ToInt32();
-            }
-
-            else
-            {
-                checkFunctionUpdateAddress = loadConfigDirectoryAddress + Marshal.OffsetOf<ImageLoadConfigDirectory64>("GuardCFCheckFunctionPointer").ToInt32();
-            }
-
-            _processContext.Process.WriteStruct(checkFunctionUpdateAddress, checkFunctionAddress);
-
-            // Update the dispatch function pointer
-
-            if (_peImage.Headers.PEHeader!.Magic == PEMagic.PE32)
-            {
-                return;
-            }
-
-            var dispatchFunctionUpdateAddress = loadConfigDirectoryAddress + Marshal.OffsetOf<ImageLoadConfigDirectory64>("GuardCFDispatchFunctionPointer").ToInt32();
-            _processContext.Process.WriteStruct(dispatchFunctionUpdateAddress, dispatchFunctionAddress);
+            _processContext.Process.WriteStruct(loadConfigDirectoryAddress, loadConfigDirectory with { GuardCFCheckFunctionPointer = _processContext.GetNtdllSymbolAddress(checkFunctionName).ToInt32() });
         }
 
-        private void InitialiseSecurityCookie()
+        else
         {
-            var loadConfigData = _peImage.LoadConfigDirectory.GetLoadConfigData();
+            var loadConfigDirectory = _processContext.Process.ReadStruct<ImageLoadConfigDirectory64>(loadConfigDirectoryAddress);
 
-            if (loadConfigData?.SecurityCookie is null || loadConfigData.GuardFlags.HasFlag(GuardFlags.SecurityCookieUnused))
-            {
-                return;
-            }
+            // Update the check and dispatch function pointers
 
-            // Generate a randomised security cookie
+            _processContext.Process.WriteStruct(loadConfigDirectoryAddress, loadConfigDirectory with { GuardCFCheckFunctionPointer = _processContext.GetNtdllSymbolAddress(checkFunctionName).ToInt64(), GuardCFDispatchFunctionPointer = _processContext.GetNtdllSymbolAddress(dispatchFunctionName).ToInt64() });
+        }
+    }
 
-            var securityCookieBytes = _peImage.Headers.PEHeader!.Magic == PEMagic.PE32 ? stackalloc byte[4] : stackalloc byte[6];
-            RandomNumberGenerator.Fill(securityCookieBytes);
+    private void InitialiseSecurityCookie()
+    {
+        var loadConfigData = _peImage.LoadConfigDirectory.GetLoadConfigData();
 
-            // Ensure the default security cookie was not generated
-
-            if (securityCookieBytes.SequenceEqual(stackalloc byte[] {0xBB, 0x40, 0xE6, 0x4E}) || securityCookieBytes.SequenceEqual(stackalloc byte[] {0x2B, 0x99, 0x2D, 0xDF, 0xA2, 0x32}))
-            {
-                securityCookieBytes[^1] += 1;
-            }
-
-            // Initialise the security cookie
-
-            var securityCookieAddress = DllBaseAddress + loadConfigData.SecurityCookie.RelativeAddress;
-            _processContext.Process.WriteSpan(securityCookieAddress, securityCookieBytes);
+        if (loadConfigData?.SecurityCookie is null || loadConfigData.GuardFlags.HasFlag(GuardFlags.SecurityCookieUnused))
+        {
+            return;
         }
 
-        private void InitialiseTlsData()
+        // Generate a randomised security cookie
+
+        var securityCookieBytes = _peImage.Headers.PEHeader!.Magic == PEMagic.PE32 ? stackalloc byte[4] : stackalloc byte[6];
+        RandomNumberGenerator.Fill(securityCookieBytes);
+
+        // Ensure the default security cookie was not generated
+
+        if (securityCookieBytes.SequenceEqual(stackalloc byte[] { 0xBB, 0x40, 0xE6, 0x4E }) || securityCookieBytes.SequenceEqual(stackalloc byte[] { 0x2B, 0x99, 0x2D, 0xDF, 0xA2, 0x32 }))
         {
-            if (_peImage.Headers.PEHeader!.ThreadLocalStorageTableDirectory.Size == 0)
-            {
-                return;
-            }
+            securityCookieBytes[^1] += 1;
+        }
 
-            var tlsDirectoryAddress = DllBaseAddress + _peImage.Headers.PEHeader!.ThreadLocalStorageTableDirectory.RelativeVirtualAddress;
-            var tlsListAddress = _processContext.GetNtdllSymbolAddress("LdrpTlsList");
-            using var pebLock = new SafePebLock(_processContext);
+        // Initialise the security cookie
 
+        var securityCookieAddress = DllBaseAddress + loadConfigData.SecurityCookie.RelativeAddress;
+        _processContext.Process.WriteSpan(securityCookieAddress, securityCookieBytes);
+    }
+
+    private void InitialiseTlsData()
+    {
+        if (_peImage.Headers.PEHeader!.ThreadLocalStorageTableDirectory.Size == 0)
+        {
+            return;
+        }
+
+        var tlsDirectoryAddress = DllBaseAddress + _peImage.Headers.PEHeader!.ThreadLocalStorageTableDirectory.RelativeVirtualAddress;
+        var tlsListAddress = _processContext.GetNtdllSymbolAddress("LdrpTlsList");
+        _processContext.PebLock.Acquire();
+
+        try
+        {
             if (_processContext.Architecture == Architecture.X86)
             {
                 // Read the TLS directory
@@ -468,8 +459,8 @@ namespace Lunar
 
                 // Write a TLS entry for the DLL into the process
 
-                _tlsData.EntryAddress = _processContext.Process.AllocateBuffer(Unsafe.SizeOf<LdrpTlsEntry32>(), ProtectionType.ReadWrite);
-                var tlsEntry = new LdrpTlsEntry32(new ListEntry32(tlsListAddress.ToInt32(), tlsListHead.Blink), tlsDirectory, _tlsData.Index);
+                _tlsData.EntryAddress = _processContext.HeapManager.AllocateBuffer(Unsafe.SizeOf<LdrpTlsEntry32>());
+                var tlsEntry = new LdrpTlsEntry32(tlsListHead with { Flink = tlsListAddress.ToInt32() }, tlsDirectory, _tlsData.Index);
 
                 try
                 {
@@ -479,28 +470,16 @@ namespace Lunar
 
                     if (tlsListAddress == tlsListTailAddress)
                     {
-                        tlsListHead = new ListEntry32(_tlsData.EntryAddress.ToInt32(), _tlsData.EntryAddress.ToInt32());
-                        _processContext.Process.WriteStruct(tlsListAddress, tlsListHead);
+                        _processContext.Process.WriteStruct(tlsListAddress, new ListEntry32(_tlsData.EntryAddress.ToInt32(), _tlsData.EntryAddress.ToInt32()));
                     }
 
                     else
                     {
+                        _processContext.Process.WriteStruct(tlsListAddress, tlsListHead with { Blink = _tlsData.EntryAddress.ToInt32() });
+
                         try
                         {
-                            var newTlsListHead = new ListEntry32(tlsListHead.Flink, _tlsData.EntryAddress.ToInt32());
-                            _processContext.Process.WriteStruct(tlsListAddress, newTlsListHead);
-
-                            try
-                            {
-                                var newTlsListTail = new ListEntry32(_tlsData.EntryAddress.ToInt32(), tlsListTail.Blink);
-                                _processContext.Process.WriteStruct(tlsListTailAddress, newTlsListTail);
-                            }
-
-                            catch
-                            {
-                                Executor.IgnoreExceptions(() => _processContext.Process.WriteStruct(_tlsData.EntryAddress, tlsListHead));
-                                throw;
-                            }
+                            _processContext.Process.WriteStruct(tlsListTailAddress, tlsListTail with { Flink = _tlsData.EntryAddress.ToInt32() });
                         }
 
                         catch
@@ -513,7 +492,7 @@ namespace Lunar
 
                 catch
                 {
-                    Executor.IgnoreExceptions(() => _processContext.Process.FreeBuffer(_tlsData.EntryAddress));
+                    Executor.IgnoreExceptions(() => _processContext.HeapManager.FreeBuffer(_tlsData.EntryAddress));
                     throw;
                 }
             }
@@ -535,10 +514,10 @@ namespace Lunar
                 var tlsListTailAddress = UnsafeHelpers.WrapPointer(tlsListHead.Blink);
                 var tlsListTail = _processContext.Process.ReadStruct<ListEntry64>(tlsListTailAddress);
 
-                // Write a TLS entry for the DLL into the process
+                // Write a TLS entry for the DLL into the process heap
 
-                _tlsData.EntryAddress = _processContext.Process.AllocateBuffer(Unsafe.SizeOf<LdrpTlsEntry64>(), ProtectionType.ReadWrite);
-                var tlsEntry = new LdrpTlsEntry64(new ListEntry64(tlsListAddress.ToInt64(), tlsListHead.Blink), tlsDirectory, _tlsData.Index);
+                _tlsData.EntryAddress = _processContext.HeapManager.AllocateBuffer(Unsafe.SizeOf<LdrpTlsEntry64>());
+                var tlsEntry = new LdrpTlsEntry64(tlsListHead with { Flink = tlsListAddress.ToInt64() }, tlsDirectory, _tlsData.Index);
 
                 try
                 {
@@ -548,28 +527,16 @@ namespace Lunar
 
                     if (tlsListAddress == tlsListTailAddress)
                     {
-                        tlsListHead = new ListEntry64(_tlsData.EntryAddress.ToInt64(), _tlsData.EntryAddress.ToInt64());
-                        _processContext.Process.WriteStruct(tlsListAddress, tlsListHead);
+                        _processContext.Process.WriteStruct(tlsListAddress, new ListEntry64(_tlsData.EntryAddress.ToInt64(), _tlsData.EntryAddress.ToInt64()));
                     }
 
                     else
                     {
+                        _processContext.Process.WriteStruct(tlsListAddress, tlsListHead with { Blink = _tlsData.EntryAddress.ToInt64() });
+
                         try
                         {
-                            var newTlsListHead = new ListEntry64(tlsListHead.Flink, _tlsData.EntryAddress.ToInt64());
-                            _processContext.Process.WriteStruct(tlsListAddress, newTlsListHead);
-
-                            try
-                            {
-                                var newTlsListTail = new ListEntry64(_tlsData.EntryAddress.ToInt64(), tlsListTail.Blink);
-                                _processContext.Process.WriteStruct(tlsListTailAddress, newTlsListTail);
-                            }
-
-                            catch
-                            {
-                                Executor.IgnoreExceptions(() => _processContext.Process.WriteStruct(_tlsData.EntryAddress, tlsListHead));
-                                throw;
-                            }
+                            _processContext.Process.WriteStruct(tlsListTailAddress, tlsListTail with { Flink = _tlsData.EntryAddress.ToInt64() });
                         }
 
                         catch
@@ -582,311 +549,335 @@ namespace Lunar
 
                 catch
                 {
-                    Executor.IgnoreExceptions(() => _processContext.Process.FreeBuffer(_tlsData.EntryAddress));
+                    Executor.IgnoreExceptions(() => _processContext.HeapManager.FreeBuffer(_tlsData.EntryAddress));
                     throw;
                 }
             }
         }
 
-        private void InsertExceptionHandlers()
+        finally
         {
-            var functionTableAddress = _processContext.GetNtdllSymbolAddress("LdrpInvertedFunctionTable");
-            using var pebLock = new SafePebLock(_processContext);
+            Executor.IgnoreExceptions(() => _processContext.PebLock.Release());
+        }
+    }
 
+    private void InsertExceptionHandlers()
+    {
+        var functionTableAddress = _processContext.GetNtdllSymbolAddress("LdrpInvertedFunctionTable");
+        _processContext.PebLock.Acquire();
+
+        try
+        {
             // Read the function table
 
             var functionTable = _processContext.Process.ReadStruct<InvertedFunctionTable>(functionTableAddress);
 
-            if (functionTable.Overflow == 1)
+            if (functionTable.Overflow)
             {
                 return;
             }
 
-            if (_peImage.Headers.PEHeader!.Magic == PEMagic.PE32)
+            if (functionTable.CurrentSize == functionTable.MaximumSize)
             {
-                var loadConfigData = _peImage.LoadConfigDirectory.GetLoadConfigData();
+                // Mark the function table as overflowed
 
-                if (loadConfigData is null)
-                {
-                    return;
-                }
-
-                // Read the function table entry list
-
-                var functionTableEntryListAddress = functionTableAddress + Unsafe.SizeOf<InvertedFunctionTable>();
-                var functionTableEntryList = _processContext.Process.ReadSpan<InvertedFunctionTableEntry32>(functionTableEntryListAddress, Constants.InvertedFunctionTableSize);
-
-                // Find the index where the entry for the DLL should be inserted
-
-                var insertionIndex = 1;
-
-                while (insertionIndex < functionTable.Count)
-                {
-                    if ((uint) DllBaseAddress.ToInt32() < (uint) functionTableEntryList[insertionIndex].ImageBase)
-                    {
-                        break;
-                    }
-
-                    insertionIndex += 1;
-                }
-
-                if (insertionIndex < functionTable.Count)
-                {
-                    // Shift the existing elements to make space for the entry for the DLL
-
-                    for (var entryIndex = functionTable.Count - 1; entryIndex >= insertionIndex; entryIndex -= 1)
-                    {
-                        functionTableEntryList[entryIndex + 1] = functionTableEntryList[entryIndex];
-                    }
-                }
-
-                // Read the shared user data
-
-                var sharedUserDataAddress = UnsafeHelpers.WrapPointer(Constants.SharedUserDataAddress);
-                var sharedUserData = _processContext.Process.ReadStruct<KUserSharedData>(sharedUserDataAddress);
-
-                // Encode the address of the exception directory using the system pointer encoding algorithm
-
-                var exceptionDirectoryAddress = DllBaseAddress + loadConfigData.ExceptionTable!.RelativeAddress;
-                var xoredAddress = (uint) exceptionDirectoryAddress.ToInt32() ^ (uint) sharedUserData.Cookie;
-                var lowerCookieBits = sharedUserData.Cookie & 0x1F;
-                var rotatedAddress = (xoredAddress >> lowerCookieBits) | (xoredAddress << (32 - lowerCookieBits));
-
-                // Update the function table entry list
-
-                functionTableEntryList[insertionIndex] = new InvertedFunctionTableEntry32((int) rotatedAddress, DllBaseAddress.ToInt32(), _peImage.Headers.PEHeader!.SizeOfImage, loadConfigData.ExceptionTable!.HandlerCount);
-                _processContext.Process.WriteSpan(functionTableEntryListAddress, functionTableEntryList);
+                functionTable = functionTable with { Overflow = true };
+                _processContext.Process.WriteStruct(functionTableAddress, functionTable);
             }
 
             else
             {
-                // Read the function table entry list
-
-                var functionTableEntryListAddress = functionTableAddress + Unsafe.SizeOf<InvertedFunctionTable>();
-                var functionTableEntryList = _processContext.Process.ReadSpan<InvertedFunctionTableEntry64>(functionTableEntryListAddress, Constants.InvertedFunctionTableSize);
-
-                // Find the index where the entry for the DLL should be inserted
-
-                var insertionIndex = 1;
-
-                while (insertionIndex < functionTable.Count)
+                if (_peImage.Headers.PEHeader!.Magic == PEMagic.PE32)
                 {
-                    if ((ulong) DllBaseAddress.ToInt64() < (ulong) functionTableEntryList[insertionIndex].ImageBase)
+                    var loadConfigData = _peImage.LoadConfigDirectory.GetLoadConfigData();
+
+                    if (loadConfigData is null)
                     {
-                        break;
+                        return;
                     }
 
-                    insertionIndex += 1;
-                }
+                    // Read the function table entry list
 
-                if (insertionIndex < functionTable.Count)
-                {
-                    // Shift the existing elements to make space for the entry for the DLL
+                    var functionTableEntryListAddress = functionTableAddress + Unsafe.SizeOf<InvertedFunctionTable>();
+                    var functionTableEntryList = _processContext.Process.ReadSpan<InvertedFunctionTableEntry32>(functionTableEntryListAddress, Constants.InvertedFunctionTableSize);
 
-                    for (var entryIndex = functionTable.Count - 1; entryIndex >= insertionIndex; entryIndex -= 1)
+                    // Find the index where the entry for the DLL should be inserted
+
+                    var insertionIndex = 1;
+
+                    while (insertionIndex < functionTable.CurrentSize)
                     {
-                        functionTableEntryList[entryIndex + 1] = functionTableEntryList[entryIndex];
-                    }
-                }
+                        if ((uint) DllBaseAddress.ToInt32() < (uint) functionTableEntryList[insertionIndex].ImageBase)
+                        {
+                            break;
+                        }
 
-                // Update the function table entry list
-
-                var exceptionDirectoryAddress = DllBaseAddress + _peImage.Headers.PEHeader!.ExceptionTableDirectory.RelativeVirtualAddress;
-                functionTableEntryList[insertionIndex] = new InvertedFunctionTableEntry64(exceptionDirectoryAddress.ToInt64(), DllBaseAddress.ToInt64(), _peImage.Headers.PEHeader!.SizeOfImage, _peImage.Headers.PEHeader!.ExceptionTableDirectory.Size);
-                _processContext.Process.WriteSpan(functionTableEntryListAddress, functionTableEntryList);
-            }
-
-            // Update the function table
-
-            var overflow = functionTable.Count + 1 == functionTable.MaxCount ? 1 : 0;
-            functionTable = new InvertedFunctionTable(functionTable.Count + 1, functionTable.MaxCount, overflow);
-            _processContext.Process.WriteStruct(functionTableAddress, functionTable);
-        }
-
-        private void LoadDependencies()
-        {
-            var activationContext = new ActivationContext(_peImage.ResourceDirectory.GetManifest(), _processContext.Architecture);
-
-            foreach (var (_, dependencyName) in _peImage.ImportDirectory.GetImportDescriptors())
-            {
-                // Write the dependency file path into the process
-
-                var dependencyFilePath = _fileResolver.ResolveFilePath(_processContext.ResolveModuleName(dependencyName), activationContext);
-
-                if (dependencyFilePath is null)
-                {
-                    throw new FileNotFoundException($"Failed to resolve the dependency file path for {dependencyName}");
-                }
-
-                var dependencyFilePathAddress = _processContext.Process.AllocateBuffer(Encoding.Unicode.GetByteCount(dependencyFilePath), ProtectionType.ReadOnly);
-
-                try
-                {
-                    _processContext.Process.WriteString(dependencyFilePathAddress, dependencyFilePath);
-
-                    // Load the dependency using the Windows loader
-
-                    var dependencyAddress = _processContext.CallRoutine<IntPtr>(_processContext.GetFunctionAddress("kernel32.dll", "LoadLibraryW"), dependencyFilePathAddress);
-
-                    if (dependencyAddress == IntPtr.Zero)
-                    {
-                        throw new ApplicationException($"Failed to load the dependency {dependencyName} into the process");
+                        insertionIndex += 1;
                     }
 
-                    _processContext.NotifyModuleLoad(dependencyAddress, dependencyFilePath);
-                }
-
-                finally
-                {
-                    Executor.IgnoreExceptions(() => _processContext.Process.FreeBuffer(dependencyFilePathAddress));
-                }
-            }
-        }
-
-        private void MapHeaders()
-        {
-            var headerBytes = _dllBytes.Span[.._peImage.Headers.PEHeader!.SizeOfHeaders];
-            _processContext.Process.WriteSpan(DllBaseAddress, headerBytes);
-        }
-
-        private void MapSections()
-        {
-            foreach (var sectionHeader in _peImage.Headers.SectionHeaders.Where(sectionHeader => !sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemDiscardable)))
-            {
-                if (sectionHeader.PointerToRawData == 0 || sectionHeader.SizeOfRawData == 0 && sectionHeader.VirtualSize == 0)
-                {
-                    continue;
-                }
-
-                // Calculate the raw section size
-
-                var sectionSize = sectionHeader.SizeOfRawData;
-
-                if (sectionHeader.SizeOfRawData > sectionHeader.VirtualSize)
-                {
-                    sectionSize = sectionHeader.VirtualSize;
-                }
-
-                // Map the raw section
-
-                var sectionAddress = DllBaseAddress + sectionHeader.VirtualAddress;
-                var sectionBytes = _dllBytes.Span.Slice(sectionHeader.PointerToRawData, sectionSize);
-                _processContext.Process.WriteSpan(sectionAddress, sectionBytes);
-
-                // Determine the protection to apply to the section
-
-                ProtectionType sectionProtection;
-
-                if (sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemExecute))
-                {
-                    if (sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemWrite))
+                    if (insertionIndex < functionTable.CurrentSize)
                     {
-                        sectionProtection = sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemRead) ? ProtectionType.ExecuteReadWrite : ProtectionType.ExecuteWriteCopy;
+                        // Shift the existing elements to make space for the entry for the DLL
+
+                        for (var entryIndex = functionTable.CurrentSize - 1; entryIndex >= insertionIndex; entryIndex -= 1)
+                        {
+                            functionTableEntryList[entryIndex + 1] = functionTableEntryList[entryIndex];
+                        }
                     }
 
-                    else
-                    {
-                        sectionProtection = sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemRead) ? ProtectionType.ExecuteRead : ProtectionType.Execute;
-                    }
-                }
+                    // Read the shared user data
 
-                else if (sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemWrite))
-                {
-                    sectionProtection = sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemRead) ? ProtectionType.ReadWrite : ProtectionType.WriteCopy;
+                    var sharedUserDataAddress = UnsafeHelpers.WrapPointer(Constants.SharedUserDataAddress);
+                    var sharedUserData = _processContext.Process.ReadStruct<KUserSharedData>(sharedUserDataAddress);
+
+                    // Encode the address of the exception directory using the system pointer encoding algorithm
+
+                    var exceptionDirectoryAddress = DllBaseAddress + loadConfigData.ExceptionTable!.RelativeAddress;
+                    var rotateValue = sharedUserData.Cookie & 0x1F;
+                    var encodedAddress = BitOperations.RotateRight((uint) exceptionDirectoryAddress.ToInt32() ^ (uint) sharedUserData.Cookie, rotateValue);
+
+                    // Update the function table entry list
+
+                    functionTableEntryList[insertionIndex] = new InvertedFunctionTableEntry32((int) encodedAddress, DllBaseAddress.ToInt32(), _peImage.Headers.PEHeader!.SizeOfImage, loadConfigData.ExceptionTable!.HandlerCount);
+                    _processContext.Process.WriteSpan(functionTableEntryListAddress, functionTableEntryList);
                 }
 
                 else
                 {
-                    sectionProtection = sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemRead) ? ProtectionType.ReadOnly : ProtectionType.NoAccess;
-                }
+                    // Read the function table entry list
 
-                if (sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemNotCached))
-                {
-                    sectionProtection |= ProtectionType.NoCache;
-                }
+                    var functionTableEntryListAddress = functionTableAddress + Unsafe.SizeOf<InvertedFunctionTable>();
+                    var functionTableEntryList = _processContext.Process.ReadSpan<InvertedFunctionTableEntry64>(functionTableEntryListAddress, Constants.InvertedFunctionTableSize);
 
-                // Calculate the aligned section size
+                    // Find the index where the entry for the DLL should be inserted
 
-                var sectionAlignment = _peImage.Headers.PEHeader!.SectionAlignment;
-                var alignedSectionSize = Math.Max(sectionHeader.SizeOfRawData, sectionHeader.VirtualSize);
-                alignedSectionSize = alignedSectionSize + sectionAlignment - 1 - (alignedSectionSize + sectionAlignment - 1) % sectionAlignment;
+                    var insertionIndex = 1;
 
-                // Adjust the protection of the aligned section
-
-                _processContext.Process.ProtectBuffer(sectionAddress, alignedSectionSize, sectionProtection);
-            }
-        }
-
-        private void ReleaseTlsIndex()
-        {
-            if (_peImage.Headers.PEHeader!.ThreadLocalStorageTableDirectory.Size == 0)
-            {
-                return;
-            }
-
-            var tlsBitmapAddress = _processContext.GetNtdllSymbolAddress("LdrpTlsBitmap");
-            using var pebLock = new SafePebLock(_processContext);
-
-            // Clear the index from the TLS bitmap
-
-            _processContext.CallRoutine(_processContext.GetFunctionAddress("ntdll.dll", "RtlClearBit"), tlsBitmapAddress, _tlsData.Index);
-        }
-
-        private void RelocateImage()
-        {
-            if (_peImage.Headers.PEHeader!.Magic == PEMagic.PE32)
-            {
-                // Calculate the delta from the preferred base address
-
-                var delta = (uint) DllBaseAddress.ToInt32() - (uint) _peImage.Headers.PEHeader!.ImageBase;
-
-                Parallel.ForEach(_peImage.RelocationDirectory.GetRelocations(), relocation =>
-                {
-                    if (relocation.Type != RelocationType.HighLow)
+                    while (insertionIndex < functionTable.CurrentSize)
                     {
-                        return;
+                        if ((ulong) DllBaseAddress.ToInt64() < (ulong) functionTableEntryList[insertionIndex].ImageBase)
+                        {
+                            break;
+                        }
+
+                        insertionIndex += 1;
                     }
 
-                    // Perform the relocation
+                    if (insertionIndex < functionTable.CurrentSize)
+                    {
+                        // Shift the existing elements to make space for the entry for the DLL
 
-                    var relocationValue = MemoryMarshal.Read<uint>(_dllBytes.Span[relocation.Offset..]) + delta;
-                    MemoryMarshal.Write(_dllBytes.Span[relocation.Offset..], ref relocationValue);
-                });
+                        for (var entryIndex = functionTable.CurrentSize - 1; entryIndex >= insertionIndex; entryIndex -= 1)
+                        {
+                            functionTableEntryList[entryIndex + 1] = functionTableEntryList[entryIndex];
+                        }
+                    }
+
+                    // Update the function table entry list
+
+                    var exceptionDirectoryAddress = DllBaseAddress + _peImage.Headers.PEHeader!.ExceptionTableDirectory.RelativeVirtualAddress;
+                    functionTableEntryList[insertionIndex] = new InvertedFunctionTableEntry64(exceptionDirectoryAddress.ToInt64(), DllBaseAddress.ToInt64(), _peImage.Headers.PEHeader!.SizeOfImage, _peImage.Headers.PEHeader!.ExceptionTableDirectory.Size);
+                    _processContext.Process.WriteSpan(functionTableEntryListAddress, functionTableEntryList);
+                }
+
+                // Update the function table size
+
+                functionTable = functionTable with { CurrentSize = functionTable.CurrentSize + 1 };
+                _processContext.Process.WriteStruct(functionTableAddress, functionTable);
+            }
+        }
+
+        finally
+        {
+            Executor.IgnoreExceptions(() => _processContext.PebLock.Release());
+        }
+    }
+
+    private void LoadDependencies()
+    {
+        var activationContext = new ActivationContext(_peImage.ResourceDirectory.GetManifest(), _processContext.Architecture);
+
+        foreach (var (_, dependencyName) in _peImage.ImportDirectory.GetImportDescriptors())
+        {
+            // Write the dependency file path into the process
+
+            var dependencyFilePath = _fileResolver.ResolveFilePath(_processContext.ResolveModuleName(dependencyName, null), activationContext);
+
+            if (dependencyFilePath is null)
+            {
+                throw new FileNotFoundException($"Failed to resolve the dependency file path for {dependencyName}");
+            }
+
+            var dependencyFilePathAddress = _processContext.Process.AllocateBuffer(Encoding.Unicode.GetByteCount(dependencyFilePath), ProtectionType.ReadOnly);
+
+            try
+            {
+                _processContext.Process.WriteString(dependencyFilePathAddress, dependencyFilePath);
+
+                // Load the dependency using the Windows loader
+
+                var dependencyAddress = _processContext.CallRoutine<IntPtr>(_processContext.GetFunctionAddress("kernel32.dll", "LoadLibraryW"), dependencyFilePathAddress);
+
+                if (dependencyAddress == IntPtr.Zero)
+                {
+                    throw new ApplicationException($"Failed to load the dependency {dependencyName} into the process");
+                }
+
+                _processContext.RecordModuleLoad(dependencyAddress, dependencyFilePath);
+            }
+
+            finally
+            {
+                Executor.IgnoreExceptions(() => _processContext.Process.FreeBuffer(dependencyFilePathAddress));
+            }
+        }
+    }
+
+    private void MapHeaders()
+    {
+        var headerBytes = _dllBytes.Span[.._peImage.Headers.PEHeader!.SizeOfHeaders];
+        _processContext.Process.WriteSpan(DllBaseAddress, headerBytes);
+    }
+
+    private void MapSections()
+    {
+        foreach (var sectionHeader in _peImage.Headers.SectionHeaders.Where(sectionHeader => !sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemDiscardable)))
+        {
+            if (sectionHeader.PointerToRawData == 0 || sectionHeader.SizeOfRawData == 0 && sectionHeader.VirtualSize == 0)
+            {
+                continue;
+            }
+
+            // Calculate the raw section size
+
+            var sectionSize = sectionHeader.SizeOfRawData;
+
+            if (sectionHeader.SizeOfRawData > sectionHeader.VirtualSize)
+            {
+                sectionSize = sectionHeader.VirtualSize;
+            }
+
+            // Map the raw section
+
+            var sectionAddress = DllBaseAddress + sectionHeader.VirtualAddress;
+            var sectionBytes = _dllBytes.Span.Slice(sectionHeader.PointerToRawData, sectionSize);
+            _processContext.Process.WriteSpan(sectionAddress, sectionBytes);
+
+            // Determine the protection to apply to the section
+
+            ProtectionType sectionProtection;
+
+            if (sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemExecute))
+            {
+                if (sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemWrite))
+                {
+                    sectionProtection = sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemRead) ? ProtectionType.ExecuteReadWrite : ProtectionType.ExecuteWriteCopy;
+                }
+
+                else
+                {
+                    sectionProtection = sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemRead) ? ProtectionType.ExecuteRead : ProtectionType.Execute;
+                }
+            }
+
+            else if (sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemWrite))
+            {
+                sectionProtection = sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemRead) ? ProtectionType.ReadWrite : ProtectionType.WriteCopy;
             }
 
             else
             {
-                // Calculate the delta from the preferred base address
-
-                var delta = (ulong) DllBaseAddress.ToInt64() - _peImage.Headers.PEHeader!.ImageBase;
-
-                Parallel.ForEach(_peImage.RelocationDirectory.GetRelocations(), relocation =>
-                {
-                    if (relocation.Type != RelocationType.Dir64)
-                    {
-                        return;
-                    }
-
-                    // Perform the relocation
-
-                    var relocationValue = MemoryMarshal.Read<ulong>(_dllBytes.Span[relocation.Offset..]) + delta;
-                    MemoryMarshal.Write(_dllBytes.Span[relocation.Offset..], ref relocationValue);
-                });
+                sectionProtection = sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemRead) ? ProtectionType.ReadOnly : ProtectionType.NoAccess;
             }
+
+            if (sectionHeader.SectionCharacteristics.HasFlag(SectionCharacteristics.MemNotCached))
+            {
+                sectionProtection |= ProtectionType.NoCache;
+            }
+
+            // Calculate the aligned section size
+
+            var sectionAlignment = _peImage.Headers.PEHeader!.SectionAlignment;
+            var alignedSectionSize = Math.Max(sectionHeader.SizeOfRawData, sectionHeader.VirtualSize);
+            alignedSectionSize = alignedSectionSize + sectionAlignment - 1 - (alignedSectionSize + sectionAlignment - 1) % sectionAlignment;
+
+            // Adjust the protection of the aligned section
+
+            _processContext.Process.ProtectBuffer(sectionAddress, alignedSectionSize, sectionProtection);
+        }
+    }
+
+    private void ReleaseTlsIndex()
+    {
+        if (_peImage.Headers.PEHeader!.ThreadLocalStorageTableDirectory.Size == 0)
+        {
+            return;
         }
 
-        private void ReserveTlsIndex()
+        var tlsBitmapAddress = _processContext.GetNtdllSymbolAddress("LdrpTlsBitmap");
+
+        // Clear the index from the TLS bitmap
+
+        _processContext.CallRoutine(_processContext.GetFunctionAddress("ntdll.dll", "RtlClearBit"), tlsBitmapAddress, _tlsData.Index);
+    }
+
+    private void RelocateImage()
+    {
+        if (_peImage.Headers.PEHeader!.Magic == PEMagic.PE32)
         {
-            if (_peImage.Headers.PEHeader!.ThreadLocalStorageTableDirectory.Size == 0)
+            // Calculate the delta from the preferred base address
+
+            var delta = (uint) DllBaseAddress.ToInt32() - (uint) _peImage.Headers.PEHeader!.ImageBase;
+
+            Parallel.ForEach(_peImage.RelocationDirectory.GetRelocations(), relocation =>
             {
-                return;
-            }
+                if (relocation.Type != RelocationType.HighLow)
+                {
+                    return;
+                }
 
-            var actualTlsBitmapSizeAddress = _processContext.GetNtdllSymbolAddress("LdrpActualBitmapSize");
-            var initialTlsBitmapBufferAddress = _processContext.GetNtdllSymbolAddress("LdrpStaticTlsBitmapVector");
-            var processHeapAddress = _processContext.GetHeapAddress();
-            var tlsBitmapAddress = _processContext.GetNtdllSymbolAddress("LdrpTlsBitmap");
+                // Perform the relocation
 
+                var relocationValue = MemoryMarshal.Read<uint>(_dllBytes.Span[relocation.Offset..]) + delta;
+                MemoryMarshal.Write(_dllBytes.Span[relocation.Offset..], ref relocationValue);
+            });
+        }
+
+        else
+        {
+            // Calculate the delta from the preferred base address
+
+            var delta = (ulong) DllBaseAddress.ToInt64() - _peImage.Headers.PEHeader!.ImageBase;
+
+            Parallel.ForEach(_peImage.RelocationDirectory.GetRelocations(), relocation =>
+            {
+                if (relocation.Type != RelocationType.Dir64)
+                {
+                    return;
+                }
+
+                // Perform the relocation
+
+                var relocationValue = MemoryMarshal.Read<ulong>(_dllBytes.Span[relocation.Offset..]) + delta;
+                MemoryMarshal.Write(_dllBytes.Span[relocation.Offset..], ref relocationValue);
+            });
+        }
+    }
+
+    private void ReserveTlsIndex()
+    {
+        if (_peImage.Headers.PEHeader!.ThreadLocalStorageTableDirectory.Size == 0)
+        {
+            return;
+        }
+
+        var actualTlsBitmapSizeAddress = _processContext.GetNtdllSymbolAddress("LdrpActualBitmapSize");
+        var initialTlsBitmapBufferAddress = _processContext.GetNtdllSymbolAddress("LdrpStaticTlsBitmapVector");
+        var tlsBitmapAddress = _processContext.GetNtdllSymbolAddress("LdrpTlsBitmap");
+        _processContext.PebLock.Acquire();
+
+        try
+        {
             if (_processContext.Architecture == Architecture.X86)
             {
                 // Read the TLS bitmap
@@ -895,7 +886,7 @@ namespace Lunar
 
                 if (tlsBitmap.SizeOfBitmap == 0)
                 {
-                    tlsBitmap = new RtlBitmap32(Constants.TlsBitmapSize, initialTlsBitmapBufferAddress.ToInt32());
+                    tlsBitmap = tlsBitmap with { Buffer = initialTlsBitmapBufferAddress.ToInt32(), SizeOfBitmap = Constants.TlsBitmapSize };
 
                     // Initialise the actual TLS bitmap size
 
@@ -910,6 +901,8 @@ namespace Lunar
 
                     if (_tlsData.Index != -1)
                     {
+                        _tlsData.ModifiedBitmap = false;
+
                         return;
                     }
 
@@ -922,12 +915,7 @@ namespace Lunar
                     {
                         // Allocate an extended TLS bitmap buffer in the process heap
 
-                        var newTlsBitmapBufferAddress = _processContext.CallRoutine<IntPtr>(_processContext.GetFunctionAddress("kernel32.dll", "HeapAlloc"), processHeapAddress, HeapAllocationType.ZeroMemory, actualBitmapSizeIncrement);
-
-                        if (newTlsBitmapBufferAddress == IntPtr.Zero)
-                        {
-                            throw new ApplicationException("Failed to allocate an extended TLS bitmap buffer in the process");
-                        }
+                        var newTlsBitmapBufferAddress = _processContext.HeapManager.AllocateBuffer(actualBitmapSizeIncrement);
 
                         // Copy over the current TLS bitmap buffer data
 
@@ -940,13 +928,10 @@ namespace Lunar
                         {
                             // Free the current TLS bitmap buffer
 
-                            if (!_processContext.CallRoutine<bool>(_processContext.GetFunctionAddress("kernel32.dll", "HeapFree"), processHeapAddress, 0, currentTlsBitmapBufferAddress))
-                            {
-                                throw new ApplicationException("Failed to free the old TLS bitmap buffer in the process");
-                            }
+                            _processContext.HeapManager.FreeBuffer(currentTlsBitmapBufferAddress);
                         }
 
-                        tlsBitmap = new RtlBitmap32(tlsBitmap.SizeOfBitmap + Constants.TlsBitmapSize, newTlsBitmapBufferAddress.ToInt32());
+                        tlsBitmap = tlsBitmap with { Buffer = newTlsBitmapBufferAddress.ToInt32(), SizeOfBitmap = tlsBitmap.SizeOfBitmap + Constants.TlsBitmapSize };
 
                         // Update the actual TLS bitmap size
 
@@ -955,7 +940,7 @@ namespace Lunar
 
                     else
                     {
-                        tlsBitmap = new RtlBitmap32(tlsBitmap.SizeOfBitmap + Constants.TlsBitmapSize, tlsBitmap.Buffer);
+                        tlsBitmap = tlsBitmap with { SizeOfBitmap = tlsBitmap.SizeOfBitmap + Constants.TlsBitmapSize };
                     }
                 }
 
@@ -972,7 +957,7 @@ namespace Lunar
 
                 if (tlsBitmap.SizeOfBitmap == 0)
                 {
-                    tlsBitmap = new RtlBitmap64(Constants.TlsBitmapSize, initialTlsBitmapBufferAddress.ToInt64());
+                    tlsBitmap = tlsBitmap with { Buffer = initialTlsBitmapBufferAddress.ToInt64(), SizeOfBitmap = Constants.TlsBitmapSize };
 
                     // Initialise the actual TLS bitmap size
 
@@ -987,6 +972,8 @@ namespace Lunar
 
                     if (_tlsData.Index != -1)
                     {
+                        _tlsData.ModifiedBitmap = false;
+
                         return;
                     }
 
@@ -999,12 +986,7 @@ namespace Lunar
                     {
                         // Allocate a new TLS bitmap buffer in the process heap
 
-                        var newTlsBitmapBufferAddress = _processContext.CallRoutine<IntPtr>(_processContext.GetFunctionAddress("kernel32.dll", "HeapAlloc"), processHeapAddress, HeapAllocationType.ZeroMemory, actualBitmapSizeIncrement);
-
-                        if (newTlsBitmapBufferAddress == IntPtr.Zero)
-                        {
-                            throw new ApplicationException("Failed to allocate an extended TLS bitmap buffer in the process");
-                        }
+                        var newTlsBitmapBufferAddress = _processContext.HeapManager.AllocateBuffer(actualBitmapSizeIncrement);
 
                         // Copy over the current TLS bitmap buffer data
 
@@ -1017,13 +999,10 @@ namespace Lunar
                         {
                             // Free the current TLS bitmap buffer
 
-                            if (!_processContext.CallRoutine<bool>(_processContext.GetFunctionAddress("kernel32.dll", "HeapFree"), processHeapAddress, 0, currentTlsBitmapBufferAddress))
-                            {
-                                throw new ApplicationException("Failed to free the old TLS bitmap buffer in the process");
-                            }
+                            _processContext.HeapManager.FreeBuffer(currentTlsBitmapBufferAddress);
                         }
 
-                        tlsBitmap = new RtlBitmap64(tlsBitmap.SizeOfBitmap + Constants.TlsBitmapSize, newTlsBitmapBufferAddress.ToInt64());
+                        tlsBitmap = tlsBitmap with { Buffer = newTlsBitmapBufferAddress.ToInt64(), SizeOfBitmap = tlsBitmap.SizeOfBitmap + Constants.TlsBitmapSize };
 
                         // Update the actual TLS bitmap size
 
@@ -1032,7 +1011,7 @@ namespace Lunar
 
                     else
                     {
-                        tlsBitmap = new RtlBitmap64(tlsBitmap.SizeOfBitmap + Constants.TlsBitmapSize, tlsBitmap.Buffer);
+                        tlsBitmap = tlsBitmap with { SizeOfBitmap = tlsBitmap.SizeOfBitmap + Constants.TlsBitmapSize };
                     }
                 }
 
@@ -1053,11 +1032,19 @@ namespace Lunar
             }
         }
 
-        private void RemoveExceptionHandlers()
+        finally
         {
-            var functionTableAddress = _processContext.GetNtdllSymbolAddress("LdrpInvertedFunctionTable");
-            using var pebLock = new SafePebLock(_processContext);
+            Executor.IgnoreExceptions(() => _processContext.PebLock.Release());
+        }
+    }
 
+    private void RemoveExceptionHandlers()
+    {
+        var functionTableAddress = _processContext.GetNtdllSymbolAddress("LdrpInvertedFunctionTable");
+        _processContext.PebLock.Acquire();
+
+        try
+        {
             // Read the function table
 
             var functionTable = _processContext.Process.ReadStruct<InvertedFunctionTable>(functionTableAddress);
@@ -1080,7 +1067,7 @@ namespace Lunar
 
                 var removalIndex = 1;
 
-                while (removalIndex < functionTable.Count)
+                while (removalIndex < functionTable.CurrentSize)
                 {
                     if (DllBaseAddress.ToInt32() == functionTableEntryList[removalIndex].ImageBase)
                     {
@@ -1090,11 +1077,11 @@ namespace Lunar
                     removalIndex += 1;
                 }
 
-                if (removalIndex < functionTable.Count - 1)
+                if (removalIndex < functionTable.CurrentSize - 1)
                 {
                     // Shift the existing elements to overwrite the entry for the DLL
 
-                    for (var entryIndex = removalIndex; entryIndex < functionTable.Count; entryIndex += 1)
+                    for (var entryIndex = removalIndex; entryIndex < functionTable.CurrentSize; entryIndex += 1)
                     {
                         functionTableEntryList[entryIndex] = functionTableEntryList[entryIndex + 1];
                     }
@@ -1121,7 +1108,7 @@ namespace Lunar
 
                 var removalIndex = 1;
 
-                while (removalIndex < functionTable.Count)
+                while (removalIndex < functionTable.CurrentSize)
                 {
                     if (DllBaseAddress.ToInt64() == functionTableEntryList[removalIndex].ImageBase)
                     {
@@ -1131,11 +1118,11 @@ namespace Lunar
                     removalIndex += 1;
                 }
 
-                if (removalIndex < functionTable.Count - 1)
+                if (removalIndex < functionTable.CurrentSize - 1)
                 {
                     // Shift the existing elements to overwrite the entry for the DLL
 
-                    for (var entryIndex = removalIndex; entryIndex < functionTable.Count; entryIndex += 1)
+                    for (var entryIndex = removalIndex; entryIndex < functionTable.CurrentSize; entryIndex += 1)
                     {
                         functionTableEntryList[entryIndex] = functionTableEntryList[entryIndex + 1];
                     }
@@ -1151,10 +1138,15 @@ namespace Lunar
                 _processContext.Process.WriteSpan(functionTableEntryListAddress, functionTableEntryList);
             }
 
-            // Update the function table
+            // Update the function table size
 
-            functionTable = new InvertedFunctionTable(functionTable.Count - 1, functionTable.MaxCount, 0);
+            functionTable = functionTable with { CurrentSize = functionTable.CurrentSize - 1, Overflow = false };
             _processContext.Process.WriteStruct(functionTableAddress, functionTable);
+        }
+
+        finally
+        {
+            Executor.IgnoreExceptions(() => _processContext.PebLock.Release());
         }
     }
 }
