@@ -1,16 +1,15 @@
-﻿namespace Lunar.Native
+﻿namespace Lunar.Native;
+
+internal static class Constants
 {
-    internal static class Constants
-    {
-        internal const int DllManifestId = 0x2;
-        internal const int InvertedFunctionTableSize = 0x200;
-        internal const int ManifestResourceId = 0x18;
-        internal const int MaxPath = 0x104;
-        internal const int MaxSymbolName = 0x7D0;
-        internal const int ShadowSpaceSize = 0x28;
-        internal const int SharedUserDataAddress = 0x7FFE0000;
-        internal const int TlsBitmapSize = 0x8;
-        internal const int TlsBitmapIncrement32 = 0x27;
-        internal const int TlsBitmapIncrement64 = 0x1F;
-    }
+    internal const int DllManifestId = 0x2;
+    internal const int InvertedFunctionTableSize = 0x200;
+    internal const int ManifestResourceId = 0x18;
+    internal const int MaxPath = 0x104;
+    internal const int MaxSymbolName = 0x7D0;
+    internal const int ShadowSpaceSize = 0x28;
+    internal const int SharedUserDataAddress = 0x7FFE0000;
+    internal const int TlsBitmapSize = 0x8;
+    internal const int TlsBitmapIncrement32 = 0x27;
+    internal const int TlsBitmapIncrement64 = 0x1F;
 }
