@@ -1,8 +1,7 @@
-namespace Lunar.Native.Enums
+namespace Lunar.Native.Enums;
+
+internal enum ModuleType
 {
-    internal enum ModuleType
-    {
-        X86 = 0x1,
-        X64 = 0x2
-    }
+    X86 = 0x1,
+    X64 = 0x2
 }
