@@ -1,8 +1,7 @@
-﻿namespace Lunar.Native.Enums
+﻿namespace Lunar.Native.Enums;
+
+internal enum ProcessInformationType
 {
-    internal enum ProcessInformationType
-    {
-        BasicInformation = 0x0,
-        Wow64Information = 0x1A
-    }
+    BasicInformation = 0x0,
+    Wow64Information = 0x1A
 }
