@@ -1,13 +1,6 @@
 ﻿using System.Runtime.InteropServices;
 
-namespace Lunar.Native.Structs
-{
-    [StructLayout(LayoutKind.Explicit, Size = 20)]
-    internal readonly struct ApiSetValueEntry
-    {
-        [FieldOffset(0xC)]
-        internal readonly int ValueOffset;
-        [FieldOffset(0x10)]
-        internal readonly int ValueCount;
-    }
-}
+namespace Lunar.Native.Structs;
+
+[StructLayout(LayoutKind.Explicit, Size = 20)]
+internal readonly record struct ApiSetValueEntry([field: FieldOffset(0x4)] int NameOffset, [field: FieldOffset(0x8)] int NameLength, [field: FieldOffset(0xC)] int ValueOffset, [field: FieldOffset(0x10)] int ValueCount);
