@@ -1,13 +1,6 @@
 using System.Runtime.InteropServices;
 
-namespace Lunar.Native.Structs
-{
-    [StructLayout(LayoutKind.Explicit, Size = 8)]
-    internal readonly struct ImageBaseRelocation
-    {
-        [FieldOffset(0x0)]
-        internal readonly int VirtualAddress;
-        [FieldOffset(0x4)]
-        internal readonly int SizeOfBlock;
-    }
-}
+namespace Lunar.Native.Structs;
+
+[StructLayout(LayoutKind.Explicit, Size = 8)]
+internal readonly record struct ImageBaseRelocation([field: FieldOffset(0x0)] int VirtualAddress, [field: FieldOffset(0x4)] int SizeOfBlock);
