@@ -1,15 +1,6 @@
 using System.Runtime.InteropServices;
 
-namespace Lunar.Native.Structs
-{
-    [StructLayout(LayoutKind.Explicit, Size = 20)]
-    internal readonly struct ImageImportDescriptor
-    {
-        [FieldOffset(0x0)]
-        internal readonly int OriginalFirstThunk;
-        [FieldOffset(0xC)]
-        internal readonly int Name;
-        [FieldOffset(0x10)]
-        internal readonly int FirstThunk;
-    }
-}
+namespace Lunar.Native.Structs;
+
+[StructLayout(LayoutKind.Explicit, Size = 20)]
+internal readonly record struct ImageImportDescriptor([field: FieldOffset(0x0)] int OriginalFirstThunk, [field: FieldOffset(0xC)] int Name, [field: FieldOffset(0x10)] int FirstThunk);
