@@ -1,13 +1,6 @@
 ﻿using System.Runtime.InteropServices;
 
-namespace Lunar.Native.Structs
-{
-    [StructLayout(LayoutKind.Explicit, Size = 16)]
-    internal readonly struct ImageResourceDirectory
-    {
-        [FieldOffset(0xC)]
-        internal readonly short NumberOfNameEntries;
-        [FieldOffset(0xE)]
-        internal readonly short NumberOfIdEntries;
-    }
-}
+namespace Lunar.Native.Structs;
+
+[StructLayout(LayoutKind.Explicit, Size = 16)]
+internal readonly record struct ImageResourceDirectory([field: FieldOffset(0xC)] short NumberOfNameEntries, [field: FieldOffset(0xE)] short NumberOfIdEntries);
