@@ -1,22 +1,6 @@
 ﻿using System.Runtime.InteropServices;
 
-namespace Lunar.Native.Structs
-{
-    [StructLayout(LayoutKind.Explicit, Size = 16)]
-    internal readonly struct InvertedFunctionTable
-    {
-        [FieldOffset(0x0)]
-        internal readonly int Count;
-        [FieldOffset(0x4)]
-        internal readonly int MaxCount;
-        [FieldOffset(0xC)]
-        internal readonly int Overflow;
+namespace Lunar.Native.Structs;
 
-        internal InvertedFunctionTable(int count, int maxCount, int overflow)
-        {
-            Count = count;
-            MaxCount = maxCount;
-            Overflow = overflow;
-        }
-    }
-}
+[StructLayout(LayoutKind.Explicit, Size = 16)]
+internal readonly record struct InvertedFunctionTable([field: FieldOffset(0x0)] int CurrentSize, [field: FieldOffset(0x4)] int MaximumSize, [field: FieldOffset(0xC)] bool Overflow);
