@@ -1,11 +1,6 @@
 ﻿using System.Runtime.InteropServices;
 
-namespace Lunar.Native.Structs
-{
-    [StructLayout(LayoutKind.Explicit, Size = 1808)]
-    internal readonly struct KUserSharedData
-    {
-        [FieldOffset(0x330)]
-        internal readonly int Cookie;
-    }
-}
+namespace Lunar.Native.Structs;
+
+[StructLayout(LayoutKind.Explicit, Size = 1840)]
+internal readonly record struct KUserSharedData([field: FieldOffset(0x330)] int Cookie);
