@@ -1,11 +1,6 @@
 ﻿using System.Runtime.InteropServices;
 
-namespace Lunar.Native.Structs
-{
-    [StructLayout(LayoutKind.Explicit, Size = 48)]
-    internal readonly struct ProcessBasicInformation64
-    {
-        [FieldOffset(0x8)]
-        internal readonly long PebBaseAddress;
-    }
-}
+namespace Lunar.Native.Structs;
+
+[StructLayout(LayoutKind.Explicit, Size = 48)]
+internal readonly record struct ProcessBasicInformation64([field: FieldOffset(0x8)] long PebBaseAddress);
