@@ -1,22 +1,6 @@
 ﻿using System.Runtime.InteropServices;
 
-namespace Lunar.Native.Structs
-{
-    [StructLayout(LayoutKind.Explicit, Size = 88)]
-    internal readonly struct SymbolInfo
-    {
-        [FieldOffset(0x0)]
-        private readonly int SizeOfStruct;
-        [FieldOffset(0x38)]
-        internal readonly long Address;
-        [FieldOffset(0x50)]
-        private readonly int MaxNameLen;
+namespace Lunar.Native.Structs;
 
-        internal SymbolInfo(int sizeOfStruct, long address, int maxNameLen)
-        {
-            SizeOfStruct = sizeOfStruct;
-            Address = address;
-            MaxNameLen = maxNameLen;
-        }
-    }
-}
+[StructLayout(LayoutKind.Explicit, Size = 88)]
+internal readonly record struct SymbolInfo([field: FieldOffset(0x0)] int SizeOfStruct, [field: FieldOffset(0x38)] long Address, [field: FieldOffset(0x50)] int MaxNameLen);
