@@ -1,135 +1,132 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Reflection.PortableExecutable;
+﻿using System.Reflection.PortableExecutable;
 using System.Runtime.CompilerServices;
 using System.Runtime.InteropServices;
 using System.Text;
 using Lunar.Native.Structs;
 using Lunar.PortableExecutable.Records;
 
-namespace Lunar.PortableExecutable.DataDirectories
-{
-    internal sealed class ImportDirectory : DataDirectoryBase
-    {
-        internal ImportDirectory(PEHeaders headers, Memory<byte> imageBytes) : base(headers.PEHeader!.ImportTableDirectory, headers, imageBytes) { }
+namespace Lunar.PortableExecutable.DataDirectories;
 
-        internal IEnumerable<ImportDescriptor> GetImportDescriptors()
+internal sealed class ImportDirectory : DataDirectoryBase
+{
+    internal ImportDirectory(PEHeaders headers, Memory<byte> imageBytes) : base(headers.PEHeader!.ImportTableDirectory, headers, imageBytes) { }
+
+    internal IEnumerable<ImportDescriptor> GetImportDescriptors()
+    {
+        if (!IsValid)
         {
-            if (!IsValid)
+            yield break;
+        }
+
+        for (var descriptorIndex = 0;; descriptorIndex += 1)
+        {
+            // Read the descriptor
+
+            var descriptorOffset = DirectoryOffset + Unsafe.SizeOf<ImageImportDescriptor>() * descriptorIndex;
+            var descriptor = MemoryMarshal.Read<ImageImportDescriptor>(ImageBytes.Span[descriptorOffset..]);
+
+            if (descriptor.FirstThunk == 0)
             {
-                yield break;
+                break;
             }
 
-            for (var descriptorIndex = 0;; descriptorIndex += 1)
+            // Read the descriptor name
+
+            var descriptorNameOffset = RvaToOffset(descriptor.Name);
+            var descriptorNameLength = ImageBytes.Span[descriptorNameOffset..].IndexOf(byte.MinValue);
+            var descriptorName = Encoding.UTF8.GetString(ImageBytes.Span.Slice(descriptorNameOffset, descriptorNameLength));
+
+            // Read the functions imported under the descriptor
+
+            var offsetTableOffset = RvaToOffset(descriptor.FirstThunk);
+            var thunkTableOffset = descriptor.OriginalFirstThunk == 0 ? offsetTableOffset : RvaToOffset(descriptor.OriginalFirstThunk);
+            var functions = GetImportedFunctions(offsetTableOffset, thunkTableOffset);
+
+            yield return new ImportDescriptor(functions, descriptorName);
+        }
+    }
+
+    private IEnumerable<ImportedFunction> GetImportedFunctions(int offsetTableOffset, int thunkTableOffset)
+    {
+        for (var functionIndex = 0;; functionIndex += 1)
+        {
+            if (Headers.PEHeader!.Magic == PEMagic.PE32)
             {
-                // Read the descriptor
+                var functionOffset = offsetTableOffset + sizeof(int) * functionIndex;
 
-                var descriptorOffset = DirectoryOffset + Unsafe.SizeOf<ImageImportDescriptor>() * descriptorIndex;
-                var descriptor = MemoryMarshal.Read<ImageImportDescriptor>(ImageBytes.Span[descriptorOffset..]);
+                // Read the function thunk
 
-                if (descriptor.FirstThunk == 0)
+                var functionThunkOffset = thunkTableOffset + sizeof(int) * functionIndex;
+                var functionThunk = MemoryMarshal.Read<int>(ImageBytes.Span[functionThunkOffset..]);
+
+                if (functionThunk == 0)
                 {
                     break;
                 }
 
-                // Read the descriptor name
+                // Check if the function is imported via ordinal
 
-                var descriptorNameOffset = RvaToOffset(descriptor.Name);
-                var descriptorNameLength = ImageBytes.Span[descriptorNameOffset..].IndexOf(byte.MinValue);
-                var descriptorName = Encoding.UTF8.GetString(ImageBytes.Span.Slice(descriptorNameOffset, descriptorNameLength));
-
-                // Read the functions imported under the descriptor
-
-                var offsetTableOffset = RvaToOffset(descriptor.FirstThunk);
-                var thunkTableOffset = descriptor.OriginalFirstThunk == 0 ? offsetTableOffset : RvaToOffset(descriptor.OriginalFirstThunk);
-                var functions = GetImportedFunctions(offsetTableOffset, thunkTableOffset);
-
-                yield return new ImportDescriptor(functions, descriptorName);
-            }
-        }
-
-        private IEnumerable<ImportedFunction> GetImportedFunctions(int offsetTableOffset, int thunkTableOffset)
-        {
-            for (var functionIndex = 0;; functionIndex += 1)
-            {
-                if (Headers.PEHeader!.Magic == PEMagic.PE32)
+                if ((functionThunk & int.MinValue) != 0)
                 {
-                    var functionOffset = offsetTableOffset + sizeof(int) * functionIndex;
+                    var functionOrdinal = functionThunk & ushort.MaxValue;
 
-                    // Read the function thunk
-
-                    var functionThunkOffset = thunkTableOffset + sizeof(int) * functionIndex;
-                    var functionThunk = MemoryMarshal.Read<int>(ImageBytes.Span[functionThunkOffset..]);
-
-                    if (functionThunk == 0)
-                    {
-                        break;
-                    }
-
-                    // Check if the function is imported via ordinal
-
-                    if ((functionThunk & int.MinValue) != 0)
-                    {
-                        var functionOrdinal = functionThunk & ushort.MaxValue;
-
-                        yield return new ImportedFunction(null, functionOffset, functionOrdinal);
-                    }
-
-                    else
-                    {
-                        // Read the function ordinal
-
-                        var functionOrdinalOffset = RvaToOffset(functionThunk);
-                        var functionOrdinal = MemoryMarshal.Read<short>(ImageBytes.Span[functionOrdinalOffset..]);
-
-                        // Read the function name
-
-                        var functionNameOffset = functionOrdinalOffset + sizeof(short);
-                        var functionNameLength = ImageBytes.Span[functionNameOffset..].IndexOf(byte.MinValue);
-                        var functionName = Encoding.UTF8.GetString(ImageBytes.Span.Slice(functionNameOffset, functionNameLength));
-
-                        yield return new ImportedFunction(functionName, functionOffset, functionOrdinal);
-                    }
+                    yield return new ImportedFunction(null, functionOffset, functionOrdinal);
                 }
 
                 else
                 {
-                    var functionOffset = offsetTableOffset + sizeof(long) * functionIndex;
+                    // Read the function ordinal
 
-                    // Read the function thunk
+                    var functionOrdinalOffset = RvaToOffset(functionThunk);
+                    var functionOrdinal = MemoryMarshal.Read<short>(ImageBytes.Span[functionOrdinalOffset..]);
 
-                    var functionThunkOffset = thunkTableOffset + sizeof(long) * functionIndex;
-                    var functionThunk = MemoryMarshal.Read<long>(ImageBytes.Span[functionThunkOffset..]);
+                    // Read the function name
 
-                    if (functionThunk == 0)
-                    {
-                        break;
-                    }
+                    var functionNameOffset = functionOrdinalOffset + sizeof(short);
+                    var functionNameLength = ImageBytes.Span[functionNameOffset..].IndexOf(byte.MinValue);
+                    var functionName = Encoding.UTF8.GetString(ImageBytes.Span.Slice(functionNameOffset, functionNameLength));
 
-                    // Check if the function is imported via ordinal
+                    yield return new ImportedFunction(functionName, functionOffset, functionOrdinal);
+                }
+            }
 
-                    if ((functionThunk & long.MinValue) != 0)
-                    {
-                        var functionOrdinal = functionThunk & ushort.MaxValue;
+            else
+            {
+                var functionOffset = offsetTableOffset + sizeof(long) * functionIndex;
 
-                        yield return new ImportedFunction(null, functionOffset, (int) functionOrdinal);
-                    }
+                // Read the function thunk
 
-                    else
-                    {
-                        // Read the function ordinal
+                var functionThunkOffset = thunkTableOffset + sizeof(long) * functionIndex;
+                var functionThunk = MemoryMarshal.Read<long>(ImageBytes.Span[functionThunkOffset..]);
 
-                        var functionOrdinalOffset = RvaToOffset((int) functionThunk);
-                        var functionOrdinal = MemoryMarshal.Read<short>(ImageBytes.Span[functionOrdinalOffset..]);
+                if (functionThunk == 0)
+                {
+                    break;
+                }
 
-                        // Read the function name
+                // Check if the function is imported via ordinal
 
-                        var functionNameOffset = functionOrdinalOffset + sizeof(short);
-                        var functionNameLength = ImageBytes.Span[functionNameOffset..].IndexOf(byte.MinValue);
-                        var functionName = Encoding.UTF8.GetString(ImageBytes.Span.Slice(functionNameOffset, functionNameLength));
+                if ((functionThunk & long.MinValue) != 0)
+                {
+                    var functionOrdinal = functionThunk & ushort.MaxValue;
 
-                        yield return new ImportedFunction(functionName, functionOffset, functionOrdinal);
-                    }
+                    yield return new ImportedFunction(null, functionOffset, (int) functionOrdinal);
+                }
+
+                else
+                {
+                    // Read the function ordinal
+
+                    var functionOrdinalOffset = RvaToOffset((int) functionThunk);
+                    var functionOrdinal = MemoryMarshal.Read<short>(ImageBytes.Span[functionOrdinalOffset..]);
+
+                    // Read the function name
+
+                    var functionNameOffset = functionOrdinalOffset + sizeof(short);
+                    var functionNameLength = ImageBytes.Span[functionNameOffset..].IndexOf(byte.MinValue);
+                    var functionName = Encoding.UTF8.GetString(ImageBytes.Span.Slice(functionNameOffset, functionNameLength));
+
+                    yield return new ImportedFunction(functionName, functionOffset, functionOrdinal);
                 }
             }
         }
