@@ -1,40 +1,38 @@
-﻿using System;
-using System.Reflection.PortableExecutable;
+﻿using System.Reflection.PortableExecutable;
 
-namespace Lunar.PortableExecutable
+namespace Lunar.PortableExecutable;
+
+internal abstract class DataDirectoryBase
 {
-    internal abstract class DataDirectoryBase
+    private protected int DirectoryOffset { get; }
+    private protected PEHeaders Headers { get; }
+    private protected Memory<byte> ImageBytes { get; }
+    private protected bool IsValid { get; }
+
+    private protected DataDirectoryBase(DirectoryEntry directory, PEHeaders headers, Memory<byte> imageBytes)
     {
-        private protected int DirectoryOffset { get; }
-        private protected PEHeaders Headers { get; }
-        private protected Memory<byte> ImageBytes { get; }
-        private protected bool IsValid { get; }
+        headers.TryGetDirectoryOffset(directory, out var directoryOffset);
 
-        private protected DataDirectoryBase(DirectoryEntry directory, PEHeaders headers, Memory<byte> imageBytes)
-        {
-            headers.TryGetDirectoryOffset(directory, out var directoryOffset);
+        DirectoryOffset = directoryOffset;
+        Headers = headers;
+        ImageBytes = imageBytes;
+        IsValid = directoryOffset != -1;
+    }
 
-            DirectoryOffset = directoryOffset;
-            Headers = headers;
-            ImageBytes = imageBytes;
-            IsValid = directoryOffset != -1;
-        }
+    private protected int RvaToOffset(int rva)
+    {
+        var sectionHeader = Headers.SectionHeaders[Headers.GetContainingSectionIndex(rva)];
 
-        private protected int RvaToOffset(int rva)
-        {
-            var sectionHeader = Headers.SectionHeaders[Headers.GetContainingSectionIndex(rva)];
+        return rva - sectionHeader.VirtualAddress + sectionHeader.PointerToRawData;
+    }
 
-            return rva - sectionHeader.VirtualAddress + sectionHeader.PointerToRawData;
-        }
+    private protected int VaToRva(int va)
+    {
+        return va - (int) Headers.PEHeader!.ImageBase;
+    }
 
-        private protected int VaToRva(int va)
-        {
-            return va - (int) Headers.PEHeader!.ImageBase;
-        }
-
-        private protected int VaToRva(long va)
-        {
-            return (int) (va - (long) Headers.PEHeader!.ImageBase);
-        }
+    private protected int VaToRva(long va)
+    {
+        return (int) (va - (long) Headers.PEHeader!.ImageBase);
     }
 }
