@@ -1,4 +1,3 @@
-﻿namespace Lunar.PortableExecutable.Records
-{
-    internal sealed record ExceptionData(int HandlerCount, int RelativeAddress);
-}
+﻿namespace Lunar.PortableExecutable.Records;
+
+internal sealed record ExceptionData(int HandlerCount, int RelativeAddress);
