@@ -1,4 +1,3 @@
-namespace Lunar.PortableExecutable.Records
-{
-    internal sealed record ImportedFunction(string? Name, int Offset, int Ordinal);
-}
+namespace Lunar.PortableExecutable.Records;
+
+internal sealed record ImportedFunction(string? Name, int Offset, int Ordinal);
