@@ -1,6 +1,5 @@
 ﻿using Lunar.Native.Enums;
 
-namespace Lunar.PortableExecutable.Records
-{
-    internal sealed record LoadConfigData(ExceptionData? ExceptionTable, GuardFlags GuardFlags, SecurityCookie? SecurityCookie);
-}
+namespace Lunar.PortableExecutable.Records;
+
+internal sealed record LoadConfigData(ExceptionData? ExceptionTable, GuardFlags GuardFlags, SecurityCookie? SecurityCookie);
