@@ -1,6 +1,5 @@
 using Lunar.Native.Enums;
 
-namespace Lunar.PortableExecutable.Records
-{
-    internal sealed record Relocation(int Offset, RelocationType Type);
-}
+namespace Lunar.PortableExecutable.Records;
+
+internal sealed record Relocation(int Offset, RelocationType Type);
