@@ -1,4 +1,3 @@
-﻿namespace Lunar.PortableExecutable.Records
-{
-    internal sealed record SecurityCookie(int RelativeAddress);
-}
+﻿namespace Lunar.PortableExecutable.Records;
+
+internal sealed record SecurityCookie(int RelativeAddress);
