@@ -1,7 +1,5 @@
-﻿using System;
-using Lunar.PortableExecutable;
+﻿using Lunar.PortableExecutable;
 
-namespace Lunar.Remote.Records
-{
-    internal sealed record Module(IntPtr Address, PeImage PeImage);
-}
+namespace Lunar.Remote.Records;
+
+internal sealed record Module(IntPtr Address, PeImage PeImage);
