@@ -1,4 +1,3 @@
-﻿namespace Lunar.Remote.Records
-{
-    internal sealed record Symbol(int RelativeAddress);
-}
+﻿namespace Lunar.Remote.Records;
+
+internal sealed record Symbol(int RelativeAddress);
