@@ -1,6 +1,3 @@
-using System;
+namespace Lunar.FileResolution.Records;
 
-namespace Lunar.FileResolution.Records
-{
-    internal sealed record ManifestDirectory(int Hash, string Language, string Path, Version Version);
-}
+internal sealed record ManifestDirectory(int Hash, string Language, string Path, Version Version);
