@@ -1,11 +1,8 @@
-﻿using System;
+﻿namespace Lunar.Native.Enums;
 
-namespace Lunar.Native.Enums
+[Flags]
+internal enum AllocationType
 {
-    [Flags]
-    internal enum AllocationType
-    {
-        Commit = 0x1000,
-        Reserve = 0x2000
-    }
+    Commit = 0x1000,
+    Reserve = 0x2000
 }
