@@ -1,18 +1,15 @@
-﻿using System;
+﻿namespace Lunar.Native.Enums;
 
-namespace Lunar.Native.Enums
+[Flags]
+internal enum ProtectionType
 {
-    [Flags]
-    internal enum ProtectionType
-    {
-        NoAccess = 0x1,
-        ReadOnly = 0x2,
-        ReadWrite = 0x4,
-        WriteCopy = 0x8,
-        Execute = 0x10,
-        ExecuteRead = 0x20,
-        ExecuteReadWrite = 0x40,
-        ExecuteWriteCopy = 0x80,
-        NoCache = 0x200
-    }
+    NoAccess = 0x1,
+    ReadOnly = 0x2,
+    ReadWrite = 0x4,
+    WriteCopy = 0x8,
+    Execute = 0x10,
+    ExecuteRead = 0x20,
+    ExecuteReadWrite = 0x40,
+    ExecuteWriteCopy = 0x80,
+    NoCache = 0x200
 }
