@@ -1,21 +1,6 @@
 using System.Runtime.InteropServices;
 
-namespace Lunar.Native.Structs
-{
-    [StructLayout(LayoutKind.Explicit, Size = 40)]
-    internal readonly struct ImageExportDirectory
-    {
-        [FieldOffset(0x10)]
-        internal readonly int Base;
-        [FieldOffset(0x14)]
-        internal readonly int NumberOfFunctions;
-        [FieldOffset(0x18)]
-        internal readonly int NumberOfNames;
-        [FieldOffset(0x1C)]
-        internal readonly int AddressOfFunctions;
-        [FieldOffset(0x20)]
-        internal readonly int AddressOfNames;
-        [FieldOffset(0x24)]
-        internal readonly int AddressOfNameOrdinals;
-    }
-}
+namespace Lunar.Native.Structs;
+
+[StructLayout(LayoutKind.Explicit, Size = 40)]
+internal readonly record struct ImageExportDirectory([field: FieldOffset(0x10)] int Base, [field: FieldOffset(0x14)] int NumberOfFunctions, [field: FieldOffset(0x18)] int NumberOfNames, [field: FieldOffset(0x1C)] int AddressOfFunctions, [field: FieldOffset(0x20)] int AddressOfNames, [field: FieldOffset(0x24)] int AddressOfNameOrdinals);
