@@ -1,6 +1,3 @@
-﻿using System.Collections.Generic;
+﻿namespace Lunar.PortableExecutable.Records;
 
-namespace Lunar.PortableExecutable.Records
-{
-    internal sealed record ImportDescriptor(IEnumerable<ImportedFunction> Functions, string Name);
-}
+internal sealed record ImportDescriptor(IEnumerable<ImportedFunction> Functions, string Name);
