@@ -1,7 +1,3 @@
-﻿using System;
-using System.Collections.Generic;
+﻿namespace Lunar.Shellcode.Records;
 
-namespace Lunar.Shellcode.Records
-{
-    internal sealed record CallDescriptor<T>(IntPtr Address, IList<T> Arguments, IntPtr ReturnAddress);
-}
+internal sealed record CallDescriptor<T>(IntPtr Address, IList<T> Arguments, IntPtr ReturnAddress);
